+++ src/datatypes/quote_handler.rs
@@ -14,10 +14,10 @@ pub trait QuoteHandler: AssetHandler {
     fn into_arc_dispatch(self: Arc<Self>) -> Arc<dyn AssetHandler+Send+Sync>;
 
     // insert, get, update and delete for market data sources
-    async fn insert_ticker(&self, ticker: &Ticker) -> Result<usize, DataError>;
-    async fn get_ticker_id(&self, ticker: &str) -> Option<usize>;
-    async fn insert_if_new_ticker(&self, ticker: &Ticker) -> Result<usize, DataError>;
-    async fn get_ticker_by_id(&self, id: usize) -> Result<Ticker, DataError>;
+    async fn insert_ticker(&self, ticker: &Ticker) -> Result<i32, DataError>;
+    async fn get_ticker_id(&self, ticker: &str) -> Option<i32>;
+    async fn insert_if_new_ticker(&self, ticker: &Ticker) -> Result<i32, DataError>;
+    async fn get_ticker_by_id(&self, id: i32) -> Result<Ticker, DataError>;
     async fn get_all_ticker(&self) -> Result<Vec<Ticker>, DataError>;
     async fn get_all_ticker_for_source(
         &self,
@@ -27,14 +27,14 @@ pub trait QuoteHandler: AssetHandler {
     /// Get all ticker that belong to a given asset specified by its asset ID
     async fn get_all_ticker_for_asset(
         &self,
-        asset_id: usize,
+        asset_id: i32,
     ) -> Result<Vec<Ticker>, DataError>;
 
     async fn update_ticker(&self, ticker: &Ticker) -> Result<(), DataError>;
-    async fn delete_ticker(&self, id: usize) -> Result<(), DataError>;
+    async fn delete_ticker(&self, id: i32) -> Result<(), DataError>;
 
     /// Insert, get, update and delete for market data sources
-    async fn insert_quote(&self, quote: &Quote) -> Result<usize, DataError>;
+    async fn insert_quote(&self, quote: &Quote) -> Result<i32, DataError>;
 
     /// Get the last quote in database for a specific asset name on or before the given time
     async fn get_last_fx_quote_before(
@@ -46,12 +46,12 @@ pub trait QuoteHandler: AssetHandler {
     /// Get the last quote in database for a specific asset id on or before the given time
     async fn get_last_quote_before_by_id(
         &self,
-        asset_id: usize,
+        asset_id: i32,
         time: DateTime<Local>,
     ) -> Result<(Quote, Currency), DataError>;
 
-    async fn get_all_quotes_for_ticker(&self, ticker_id: usize) -> Result<Vec<Quote>, DataError>;
+    async fn get_all_quotes_for_ticker(&self, ticker_id: i32) -> Result<Vec<Quote>, DataError>;
     async fn update_quote(&self, quote: &Quote) -> Result<(), DataError>;
-    async fn delete_quote(&self, id: usize) -> Result<(), DataError>;
+    async fn delete_quote(&self, id: i32) -> Result<(), DataError>;
     async fn remove_duplicates(&self) -> Result<(), DataError>;
 }
