+++ src/datatypes/stock.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
 pub struct Stock {
-    pub id: Option<usize>,
+    pub id: Option<i32>,
     pub name: String,
     pub isin: Option<String>,
     pub wkn: Option<String>,
@@ -13,7 +13,7 @@ pub struct Stock {
 
 impl Stock {
     pub fn new(
-        id: Option<usize>,
+        id: Option<i32>,
         name: String,
         isin: Option<String>,
         wkn: Option<String>,
@@ -31,7 +31,7 @@ impl Stock {
 
 impl DataItem for Stock {
     // get id or return error if id hasn't been set yet
-    fn get_id(&self) -> Result<usize, DataError> {
+    fn get_id(&self) -> Result<i32, DataError> {
         match self.id {
             Some(id) => Ok(id),
             None => Err(DataError::DataAccessFailure(
@@ -40,7 +40,7 @@ impl DataItem for Stock {
         }
     }
     // set id or return error if id has already been set
-    fn set_id(&mut self, id: usize) -> Result<(), DataError> {
+    fn set_id(&mut self, id: i32) -> Result<(), DataError> {
         match self.id {
             Some(_) => Err(DataError::DataAccessFailure(
                 "Can't change id of persistent stock".to_string(),
